Rules for tests on the app-level
This generally applies to various multi-micro-lib cases that we had
and shall improve those situations for all participants (the team driving
a case as well as MIR reviewers).

This is a formal writeup of the discussion we had in a recent meeting [1]
where we seemed to already have general approval.

[1]: https://new.ubottu.com/meetingology/logs/ubuntu-meeting/2021/ubuntu-meeting.2021-09-07-14.29.log.html#l-56

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>