MIR-Review: combine rules and template in one place
To avoid error prone duplication this unifies the MIR review rules with
the review template.

Thereby it follows the same style that the review-template already used
for the few intermixed rules we had.

By having them in one place close to each other we avoid people missing
one or the other by accident. It also reduces the risk to modify one,
but not the other later on.

Some rules and todo move from review to request as we always expected
it from the requesting team (e.g. to commit to testing) but due to the
way it was written before it was in the review section. Much clearer now
and based on that the reporter now also has the required template
entires to confirm he knows what he signs up for.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>