Remove duplicates in rules vs statements
In various case the rules and the templates had pretty much the same
content. For cases we expect the reporter/reviewer to be "aware" they
shall be only present in RULE sections. Vice versa for cases where we
expect the reporter/reviewer to make a statement they are at a "TODO"
entry.

In trivial cases only a TODO section is needed. And in most complex
cases the TODO entries hold the foundation for the required statements
while RULE entries only elaborate on details or alternatives while not
being expected in the final MIR bug content.

This streamlines the text flow a lot and also further reduced
redundancy.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>