Better handling of alternative answers
The review of the rework identified the need for better handling of
alternative answers. It was too easy to miss that some statements are
meant as mutually exclusive answers.

Base on a discussion we had those are now clearly marked with an
-A,-B-C,... suffix to the TODO: prefix.

Along that the guideline how to use these templates is enhanced to
explain how we expect those to be used and also got some more
clarifications in general.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>