rust/go: flexibility for refreshing vendored dependencies
Until the ecosystems settled and all projects adapted.
Ass well as the dh_* toolchain tolerates and produces reliable
go.mod/cargo.lock tracking we have to be slightly more lenient on
how exactly vendored dependencies are updated.

That shall now - for all types - be outlined in README.source.
That can for some projects still be `go mod ...` or `cargo update ...`
but in other cases might need other steps and README.source
provides the flexibility to do so.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>