Clarify to look for better alternatives
We had a few reports which could have been resolved when the reporter
evaluates the case by finding that there are better alternatives.

I hate to have people spend time in preparing for the MIR process to
then be given a Nack. Hence we should make it more clear in the
reporter section of the process that Ubuntu should avoid duplication.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>