add rereviews to the MIR process
In the last sprint a thought that has come up is the general desire
to take our mission of "quality in main" serious and that actually
re-reviews of packages already in main would need to be part of that.

This starts to document what we could do to start the discussion
as only about half the team was in the in-person session.

The rules are 60% subset of the iniital MIR rules and 10% slight
modifications or small additions. It might make more sense to not create
a second redundant template and instead somehow "mark" those
that shall be "also done" or "only done" on re-reviews in the original
template. Open for suggestions, as I said this is only meant to be a
draft to get the ball rolling.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>