Clarify helpful commands
Based on the MIR team discussion and PR feedback clarify that this
needs to be run in source and even more on built binaries.
Further clarify how to look for setuid and setgid.
And finally, those are not rules or content to leave in the final
post, so these hints are now prefixed with RULE as we usually do
for this kind of content.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>