Ask for reasoning to be in main
In most requests core Ubuntu development teams know exactly what they
need and why. Due to that it hasn't been in the rules before, but
we've seen some requesters being unclear or based on old assumptions.

Force that potential mismatch into the open by explicitly asking for
the reason what they want to achieve and allowing the MIR team to
consider that in the evaluation of the rational.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>