Drop upstream redflag for libgoa
As discussed in [1] and #54 libgoa has no reason to be excluded
other than their past integration with webkit. Therefore
goa (Gnome online accounts) itself didn't seem to be correct
to be denied on its own and was decided to be removed from
the list of libraries to watch out for.

[1]: https://ubottu.com/meetingology/logs/ubuntu-meeting/2024/ubuntu-meeting.2024-04-23-14.31.moin.txt

Fixes: #54

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>