Fix visually broken flowchart (#67)
The issue seems to be the width of text of some boxes. I used html labels to have more control over the visual results. This has the result that the renderer (for some reason) does not auto-wrap the text if an element is too wide. Apparently the text of a box does not get displayed if it is too wide.